pyproject.toml
[project]
name = "sensemaking"
version = "0.1.0"
description = "A generic response analysis platform for parsing free-text data from diverse sources"
authors = [
    {name = "Jim Moffet", email = "jim@digitalpublic.ventures"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "google-genai>=1.56.0",
    "fastapi>=0.115.0",
    "python-multipart>=0.0.12",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "xxhash>=3.4.1",
    "numpy>=1.26.0",
    "pymupdf>=1.24.0",
    "pillow>=10.4.0",
    "python-docx>=1.1.0",
    "paddleocr>=2.7.3",
    "paddlepaddle>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "black>=24.0.0",
    "pre-commit>=3.6.0",
]

[project.urls]
Homepage = "https://github.com/jimmoffet/sensemaking"
Repository = "https://github.com/jimmoffet/sensemaking.git"
Issues = "https://github.com/jimmoffet/sensemaking/issues"

[project.scripts]
cfpb-exploration = "cfpb_exploration.__main__:main"
regs-gov-exploration = "regs_dot_gov_exploration.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cfpb_exploration", "src/regs_dot_gov_exploration", "src/sensemaking_api", "src/utilities"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.ruff]
src = ["src", "tests"]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["cfpb_exploration", "regs_dot_gov_exploration", "utilities"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
